backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
joblib==1.4.2
Pillow==11.1.0
